Modified Project CMake Configuration
The previous CMake version used (3.18) was almost certainly
too strict a requirement to be useful to the vast majority
of users.

The current version we are using, 2.6, is combined with
CMake policy 48, to allow for a much wider install base.
However, policy 48 was introduced in CMake 3.0, and we had
to require this to allow for setting the project version in
the instantiation command, so it may well be possible that
CMake version 3.0 is the minimum allowable value. I am not
personally familiar with CMake enough to be able to say
whether there is some mechanism that may allow for handling
such an incompatibility in a non-fatal-error-inducing way,
so we will need to test this to be certain.